Separate rounding mode code into a module
