+++ datafusion/core/src/physical_optimizer/pipeline_checker.rs
@@ -27,8 +27,7 @@ use crate::physical_plan::{with_new_children_if_necessary, ExecutionPlan};
 use datafusion_common::config::OptimizerOptions;
 use datafusion_common::tree_node::{Transformed, TreeNode, VisitRecursion};
 use datafusion_common::DataFusionError;
-use datafusion_physical_expr::intervals::cp_solver::check_support;
-use datafusion_physical_expr::intervals::is_datatype_supported;
+use datafusion_physical_expr::intervals::{check_support, is_datatype_supported};
 use std::sync::Arc;
 
 /// The PipelineChecker rule rejects non-runnable query plans that use
